test: move hit testing service coverage to unit tests
Reclassify unit-level coverage from tests/test_containers_integration_tests/services/test_hit_testing_service.py into the unit test suite, or remove it when equivalent unit coverage already exists. Preserve real container-backed coverage and use SQLite sessions for database-dependent unit paths.
